some tiny changes

+++ testresource/__init__.py
@@ -8,14 +8,12 @@ class EchoHandler:
         return arg
 
     def channel( self, env, channel ):
-        #channel.emit( 'Hello from %s !' % channel.name )
+        channel.emit( 'Hello from %s !' % channel.name )
         ping = spawn( self._ping, channel )
-        channel.absorb( self._echo_response )
-        channel.listen()
-        ping.kill()
-        log.debug('closing ....')
+        channel.absorb( self._echo_response, channel )
+        channel.on_disconnect( ping.kill )
 
-    def _echo_response( self, channel, message ):
+    def _echo_response( self, message, channel ):
         log.debug('message received: %s' % (message,) )
         sleep(10)
         channel.emit( 'Message received: %s' % message )
@@ -33,7 +31,7 @@ class EchoHandler:
 class Worker:
     worker = None
 
-    def _handle_cmd( self, channel, message ):
+    def _handle_cmd( self, message, channel ):
         if not message == 'stop':
             channel.emit( 'ok, will do %s' % message )
             self.worker = spawn(getattr(self,message), channel )
@@ -43,12 +41,10 @@ class Worker:
 
     def __call__( self, env, channel ):
         channel.emit('gimme work !')
-        channel.absorb( self._handle_cmd )
-        channel.listen()
-        if self.worker:
-            self.worker.kill()
+        channel.absorb( self._handle_cmd, channel )
+        channel.on_disconnect( lambda: self.worker.kill() if self.worker else None )
 
-    def somework(self, channel):
+    def somework(self, channel ):
         while True:
             sleep(5)
             channel.emit( 'still working' )
